+++ isra/src/screening/screening.py
@@ -1,5 +1,13 @@
+import typer
+
 from isra.src.config.config import get_sf_values
-from isra.src.screening.screening_service import *
+from isra.src.config.constants import IR_SF_T_STRIDE, IR_SF_C_SCOPE, IR_SF_C_STANDARD_BASELINES
+from isra.src.screening.screening_service import get_all_threats, screening, get_stride_category, save_stride_category, \
+    get_attack_technique, save_attack_technique, get_attack_mitigation, save_attack_mitigation, get_all_controls, \
+    get_intended_scope, save_intended_scope, get_baseline_standard_ref, save_baseline_standard_ref, \
+    get_baseline_standard_section, save_baseline_standard_section, get_cia_triad, save_cia_triad, get_proper_cost, \
+    save_proper_cost, get_proper_cwe, save_proper_cwe, generate_question, save_question, threat_generator, \
+    control_generator, autoscreening_init, fix_component
 
 app = typer.Typer(no_args_is_help=True, add_help_option=False)
 
@@ -106,7 +114,7 @@ def question():
      depending on the given answer"""
 
     items = get_all_controls()
-    screening(items, get_question, save_question)
+    screening(items, generate_question, save_question)
 
 
 @app.command()
@@ -125,22 +133,22 @@ def new_control():
     control_generator()
 
 
-@app.command()
-def threat_screening():
-    """
-    Does all threat screenings at the same time
-    """
-    items = get_all_threats()
-    screening(items, get_complete_threat, save_complete_threat)
-
-
-@app.command()
-def control_screening():
-    """
-    Does all countermeasure screenings at the same time
-    """
-    items = get_all_controls()
-    screening(items, get_complete_control, save_complete_control)
+# @app.command()
+# def threat_screening():
+#     """
+#     Does all threat screenings at the same time
+#     """
+#     items = get_all_threats()
+#     screening(items, get_complete_threat, save_complete_threat)
+#
+#
+# @app.command()
+# def control_screening():
+#     """
+#     Does all countermeasure screenings at the same time
+#     """
+#     items = get_all_controls()
+#     screening(items, get_complete_control, save_complete_control)
 
 
 @app.command()
@@ -149,3 +157,11 @@ def autoscreening():
     Automated screening process
     """
     autoscreening_init()
+
+
+@app.command(hidden=True)
+def fix():
+    """
+    Tries to fix anything that doesn't fit the YSC schema
+    """
+    fix_component()