+++ isra/src/utils/text_functions.py
@@ -1,6 +1,7 @@
 import difflib
 import json
 import re
+from json import JSONDecodeError
 
 import itertools
 from bs4 import BeautifulSoup
@@ -24,7 +25,12 @@ def extract_json(json_string):
 
     # Parse the JSON object and return it
 
-    result = json.loads(json_object)
+    try:
+        result = json.loads(json_object)
+    except JSONDecodeError as e:
+        print("Couldn't convert JSON answer")
+        print(json_object)
+        raise e
 
     return result
 
@@ -233,29 +239,35 @@ def compare_elements(elem1, elem2, path=""):
 
 
 def set_value(variable_name, variable, value, action):
+    message = ""
+    log = True
+
     if action == "init":
         if variable == "":
             variable = value
-            print(f"Setting {value} on {variable_name} for the first time")
+            message = f"Setting '{value}' on {variable_name} for the first time"
     elif action == "replace":
         if str(variable) != str(value):
-            print(f"Replacing {variable} with {value} on {variable_name}")
+            message = f"Replacing '{variable}' with '{value}' on {variable_name}"
             variable = value
     elif action == "append":
         if isinstance(variable, list):
             if value.lower() not in [x.lower() for x in variable] and value != "":
                 variable.append(value)
-                print(f"Appending {value} to {variable_name} (list)")
+                message = f"Appending '{value}' to {variable_name} (list)"
         elif isinstance(variable, str):
             if value.lower() not in variable.lower():
                 if variable != "" and value != "":
                     variable += "||" + value
-                    print(f"Appending {value} to {variable_name} (string)")
+                    message = f"Appending '{value}' to {variable_name} (string)"
                 elif variable == "":
                     variable = value
-                    print(f"Appending {value} to {variable_name} (string)")
+                    message = f"Appending '{value}' to {variable_name} (string)"
 
         else:
             print("Type is not accepted to be appended")
+
+    if message and log:
+        print(message)
 
     return variable